Move open_fits to image/utils
