+++ vast_pipeline/pipeline/utils.py
@@ -1708,32 +1708,3 @@ def write_parquets(
     )
 
     return skyregs_df
-
-def open_fits(fits_path: Union[str, Path], memmap: Optional[bool]=True):
-    """
-    This function opens both compressed and uncompressed fits files.
-    
-    Args:
-        fits_path: Path to the fits file
-        memmap: Open the fits file with mmap.
-    
-    Returns:
-        HDUList loaded from the fits file
-    
-    Raises:
-        ValueError: File extension must be .fits or .fits.fz
-    """
-
-    if type(fits_path) == Path:
-        fits_path = str(fits_path)
-    
-    hdul = fits.open(fits_path, memmap=memmap)
-    
-    if fits_path.endswith('.fits'):
-        return hdul
-    elif fits_path.endswith('.fits.fz'):
-        return fits.HDUList(hdul[1:])
-    else:
-        raise ValueError("Unrecognised extension for {fits_path}."
-                         "File extension must be .fits or .fits.fz"
-                         )
